stock-scanner/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
apscheduler>=3.10.0
sqlalchemy>=1.4.0,<2.0.0
yfinance>=0.2.40
pykrx>=1.0.45
finance-datareader>=0.9.50
pandas>=1.5.0,<2.0.0
numpy>=1.24.0,<2.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
jinja2>=3.1.0
requests>=2.31.0
pytz>=2023.3
aiofiles>=23.1.0
lxml>=4.9.0
html5lib>=1.1
